Clean up main bot class and use service providers to load parts of the bot
